Added equipping
Crafting finally has a purpose. You can now equip items in the inventory screen with a UI preview of what's currently equipped on the home screen. Equipped items such as pickaxes affect the strength and efficiency of block breaking.

Also added options to destroy or unequip an item, as well as a description of each item.